Support Kata installation using RHCO layered image
The functionality is exposed via feature gate.

The feature gate is named as layeredImageDeployment and is managed via
the osc-feature-gates ConfigMap.  The actual MachineConfig specifying
the image details is stored in another ConfigMap:
sandboxed-containers-image-cm.

How is the layered Image Deployment supposed to work ?

- Admins create a ConfigMap with the MachineConfig embedded.
- The OSC operator reads this ConfigMap, extracts the MachineConfig, and
  creates the MachineConfig object as part of installation.
- If the feature gate is disabled or the ConfigMap is deleted, the OSC
  operator initiates a rollback/uninstallation.

Why the need to embed MachineConfig in a ConfigMap ?

If the admin directly creates the MachineConfig object then the Machine
Config Operator (MCO) will take over and start applying the
MachineConfig. We don't want that. Instead we want the OSC operator to
create the MachineConfig, so that install and rollback of Kata is
handled only by the OSC operator.

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>