podvm: Disable root disk mount points
For CoCo , we don't want root disk to be used without encryption.
Disable the root disk mounts for kata containers when
CONFIDENTIAL_COMPUTE_ENABLED is true till we have support for encrypted
disks.

Note that this means that for using container images the VM should have
sufficient memory to download and use the container images from memory.

Fixes: #KATA-3229

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>